requirements.txt
asgiref==3.8.1
Django==5.2
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-nested-admin==4.1.1
python-monkey-business==1.1.0
sqlparse==0.5.3
tzdata==2025.2
